moved deps in auth
